+++ packages/@honkit/html/src/readme.ts
@@ -7,11 +7,11 @@ import * as dom from "./dom";
     @return {Object}
 */
 function parseReadme(html) {
-    const $: any = dom.parse(html);
+    const $ = dom.parse(html);
 
     return {
         title: $("h1:first-child").text().trim(),
-        description: $("div.paragraph,p").first().text().trim(),
+        description: $("div.paragraph,p").first().text().trim()
     };
 }