+++ packages/@honkit/markup-it/syntaxes/markdown/math.js
@@ -1,6 +1,3 @@
-const ltrim = require("ltrim");
-const rtrim = require("rtrim");
-
 const reInline = require("./re/inline");
 const MarkupIt = require("../../");
 
@@ -18,8 +15,8 @@ function isInlineTex(content) {
  * @return {String}
  */
 function normalizeTeX(content, isInline) {
-    content = ltrim(content, "\n");
-    content = rtrim(content, "\n");
+    // trim new line at start and end, but preserve spaces
+    content = content.replace(/^\n+/, "").replace(/\n+$/, "");
 
     if (!isInline) {
         content = `\n${content}\n`;
@@ -27,7 +24,6 @@ function normalizeTeX(content, isInline) {
 
     return content;
 }
-
 module.exports = MarkupIt.Rule(MarkupIt.ENTITIES.MATH)
     .regExp(reInline.math, (state, match) => {
         const text = match[1];
@@ -38,8 +34,8 @@ module.exports = MarkupIt.Rule(MarkupIt.ENTITIES.MATH)
 
         return {
             data: {
-                tex: text,
-            },
+                tex: text
+            }
         };
     })
     .toText((state, token) => {
@@ -57,3 +53,4 @@ module.exports = MarkupIt.Rule(MarkupIt.ENTITIES.MATH)
 
         return output;
     });
+module.exports.normalizeTeX = normalizeTeX;