+++ packages/honkit/src/__tests__/snapshot-honkit.ts
@@ -1,9 +1,7 @@
 import path from "path";
 import { iterateDirectoryContents } from "@honkit/internal-test-utils";
 import * as bin from "../bin";
-
 it("HonKit snapshots", async () => {
-    jest.setTimeout(60 * 1000);
     const bookDir = path.join(__dirname, "__fixtures__/honkit");
     const outputDir = path.join(__dirname, "__fixtures__/honkit/_book");
     await bin.run([process.argv[0], ".", "build", bookDir, "--reload"]);
@@ -19,4 +17,4 @@ it("HonKit snapshots", async () => {
     })) {
         expect(item).toMatchSnapshot(item.filePath);
     }
-});
+}, 60 * 1000);