+++ packages/honkit/src/models/fs.ts
@@ -5,7 +5,6 @@ import File from "./file";
 import Promise from "../utils/promise";
 import error from "../utils/error";
 import PathUtil from "../utils/path";
-import { root } from "@honkit/html/lib/dom";
 
 class FS extends Immutable.Record({
     root: String(),
@@ -16,7 +15,7 @@ class FS extends Immutable.Record({
     fsReadDir: Function(),
 
     fsLoadObject: null,
-    fsReadAsStream: null,
+    fsReadAsStream: null
 }) {
     /**
      Return path to the root
@@ -52,7 +51,7 @@ class FS extends Immutable.Record({
         if (!this.isInScope(filename)) {
             throw error.FileOutOfScopeError({
                 filename: filename,
-                root: rootPath,
+                root: rootPath
             });
         }