+++ packages/honkit/src/models/output.ts
@@ -5,10 +5,10 @@ import parsePageFromString from "../parse/parsePageFromString";
 import Book from "./book";
 import LocationUtils from "../utils/location";
 import Page from "./page";
-import Logger from "../utils/logger";
+import type GeneratorState from "../output/website/state";
 
 type Assets = Immutable.List<string>;
-type State = Immutable.Map<any, any>;
+type State = GeneratorState;
 type Options = Immutable.Map<any, any>;
 class Output extends Immutable.Record({
     book: new Book(),
@@ -33,7 +33,7 @@ class Output extends Immutable.Record({
 
     // incrementalChangeFileSet for incremental building
     // If it is empty, should build all
-    incrementalChangeFileSet: Immutable.Set(),
+    incrementalChangeFileSet: Immutable.Set()
 }) {
     getBook(): Book {
         return this.get("book");
@@ -85,7 +85,7 @@ class Output extends Immutable.Record({
         }
         const newPage = parsePageFromString(page, fs.readFileSync(filePath, "utf-8"));
         return this.merge({
-            pages: pages.set(relativePath, newPage),
+            pages: pages.set(relativePath, newPage)
         });
     }