+++ packages/honkit/src/models/plugin.ts
@@ -28,7 +28,7 @@ class Plugin extends Immutable.Record(
         package: Immutable.Map(),
 
         // Content of the package itself
-        content: Immutable.Map(),
+        content: Immutable.Map()
     },
     "Plugin"
 ) {
@@ -62,11 +62,16 @@ class Plugin extends Immutable.Record(
 
     /**
      * Return the ID on NPM for this plugin
+     * return package.json's name
      * @return {string}
      */
 
     getNpmID(): string {
-        return PluginDependency.nameToNpmID(this.getName());
+        const pkg = this.getPackage().toJS();
+        if ("name" in pkg) {
+            return pkg["name"];
+        }
+        throw new Error(`${this.getName()} plugin's package.json does not have "name" fields.`);
     }
 
     /**
@@ -145,8 +150,8 @@ class Plugin extends Immutable.Record(
 
     /**
      * Create a plugin from a string
-     * @param {string}
      * @return {Plugin}
+     * @param s
      */
 
     static createFromString(s: string) {
@@ -156,7 +161,7 @@ class Plugin extends Immutable.Record(
 
         return new Plugin({
             name: name,
-            version: version || DEFAULT_VERSION,
+            version: version || DEFAULT_VERSION
         });
     }
 
@@ -168,10 +173,9 @@ class Plugin extends Immutable.Record(
     static createFromDep(dep: PluginDependency) {
         return new Plugin({
             name: dep.getName(),
-            version: dep.getVersion(),
+            version: dep.getVersion()
         });
     }
-
-    static nameToNpmID = PluginDependency.nameToNpmID;
 }
+
 export default Plugin;