+++ packages/honkit/src/plugins/listResources.ts
@@ -2,15 +2,13 @@ import Immutable from "immutable";
 import path from "path";
 import LocationUtils from "../utils/location";
 import PLUGIN_RESOURCES from "../constants/pluginResources";
+import Plugin from "../models/plugin";
 
 /**
  List all resources from a list of plugins
-
- @param {OrderedMap<String:Plugin>}
- @param {string} type
  @return {Map<String:List<{url, path}>}
  */
-function listResources(plugins, resources) {
+function listResources(plugins: Plugin[], resources) {
     return plugins.reduce((result, plugin) => {
         const npmId = plugin.getNpmID();
         const pluginResources = resources.get(plugin.getName());
@@ -24,11 +22,11 @@ function listResources(plugins, resources) {
             assets = assets.map((assetFile) => {
                 if (LocationUtils.isExternal(assetFile)) {
                     return {
-                        url: assetFile,
+                        url: assetFile
                     };
                 } else {
                     return {
-                        path: LocationUtils.normalize(path.join(npmId, assetFile)),
+                        path: LocationUtils.normalize(path.join(npmId, assetFile))
                     };
                 }
             });