+++ examples/multiple-languages/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.5](https://github.com/honkit/honkit/compare/v4.0.4...v4.0.5) (2023-03-26)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [4.0.4](https://github.com/honkit/honkit/compare/v4.0.3...v4.0.4) (2022-10-22)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [4.0.3](https://github.com/honkit/honkit/compare/v4.0.2...v4.0.3) (2022-10-22)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [4.0.2](https://github.com/honkit/honkit/compare/v4.0.1...v4.0.2) (2022-10-19)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [4.0.1](https://github.com/honkit/honkit/compare/v4.0.0...v4.0.1) (2022-10-18)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+# [4.0.0](https://github.com/honkit/honkit/compare/v3.7.5...v4.0.0) (2022-08-28)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [3.7.5](https://github.com/honkit/honkit/compare/v3.7.4...v3.7.5) (2022-07-14)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [3.7.4](https://github.com/honkit/honkit/compare/v3.7.3...v3.7.4) (2022-06-21)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [3.7.3](https://github.com/honkit/honkit/compare/v3.7.2...v3.7.3) (2022-05-28)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
+## [3.7.2](https://github.com/honkit/honkit/compare/v3.7.1...v3.7.2) (2022-05-28)
+
+**Note:** Version bump only for package @example/multiple-languages
+
+
+
+
+
 ## [3.7.1](https://github.com/honkit/honkit/compare/v3.7.0...v3.7.1) (2022-01-17)
 
 **Note:** Version bump only for package @example/multiple-languages